requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
google-genai
google-api-core
python-dotenv
pydantic
requests
tenacity
fastapi[standard]